STRINGdb-class {STRINGdb}	R Documentation

Class "STRINGdb"

Description

The R package STRINGdb provides a convenient interface to the STRING protein-protein interactions database for the R/bioconductor users. Please look at the manual/vignette to get additional informationd and examples on how to use the package. STRING is a database of known and predicted protein-protein interactions. It contains information from numerous sources, including experimental repositories, computational prediction methods and public text collections. Each interaction is associated with a combined confidence score that integrates the various evidences. STRING is regularly updated , the latest version 9.05 contains information on 5 millions proteins from more than 1100 species. The STRING web interface is freely accessible at: http://string-db.org/

Extends

All reference classes extend and inherit methods from "envRefClass".

Fields

annotations:

Object of class data.frame ~~

annotations_description:

Object of class data.frame ~~

graph:

Object of class igraph ~~

proteins:

Object of class data.frame ~~

speciesList:

Object of class data.frame ~~

species:

Object of class numeric ~~

version:

Object of class character ~~

input_directory:

Object of class character ~~

backgroundV:

Object of class vector ~~

score_threshold:

Object of class numeric ~~

Methods

set_background(background_vector):

~~

post_payload(stringIds, colors, comments, links, iframe_urls, logo_imgF, legend_imgF):

~~

plot_network(string_ids, payload_id, required_score):

~~

plot_ppi_enrichment(string_ids, file, sliceWindow, edgeWindow, windowExtendedReferenceThreshold, minVal, title):

~~

map(my_data_frame, my_data_frame_id_col_names, takeFirst, removeUnmappedRows, quiet):

~~

load():

~~

get_term_proteins(term_ids, string_ids, enableIEA):

~~

get_summary(string_ids):

~~

get_subnetwork(string_ids):

~~

get_ppi_enrichment_full(string_ids, sliceWindow, edgeWindow, windowExtendedReferenceThreshold, growingWindowLimit):

~~

get_ppi_enrichment(string_ids):

~~

get_proteins():

~~

get_png(string_ids, required_score, network_flavor, file, payload_id):

~~

get_neighbors(string_ids):

~~

get_link(string_ids, required_score, network_flavor, payload_id):

~~

get_interactions(string_ids):

~~

get_homologs_besthits(string_ids, symbets, target_species_id, bitscore_threshold):

~~

get_homologs(string_ids, target_species_id, bitscore_threshold):

~~

get_graph():

~~

get_enrichment(string_ids, category, methodMT, iea):

~~

get_clusters(string_ids, algorithm):

~~

get_annotations_desc():

~~

get_annotations():

~~

load_all():

~~

initialize(...):

~~

add_proteins_description(screen):

~~

add_diff_exp_color(screen, logFcColStr):

~~

show():

~~

Author(s)

Andrea Franceschini

References

Franceschini, A (2013). STRING v9.1: protein-protein interaction networks, with increased coverage and integration. In:'Nucleic Acids Res. 2013 Jan;41(Database issue):D808-15. doi: 10.1093/nar/gks1094. Epub 2012 Nov 29'.

See Also

http://stitch-db.org

Examples


showClass("STRINGdb")

[Package STRINGdb version 1.22.0 Index]