add_proteins_description {STRINGdb}	R Documentation

add_proteins_description

Description

Add description coluns to the proteins that are present in the data frame given in input. The data frame must contain a column named "STRING_id".

Usage


## S4 method for signature 'STRINGdb'
add_proteins_description(screen)

Arguments

screen
Dataframe containing the results of the experiment (e.g. the analyzed results of a microarray or RNAseq experiment)

Value

returns the same dataframe given in input with an additional columns containing a description of the proteins.

Author(s)

Andrea Franceschini

[Package STRINGdb version 1.22.0 Index]