get_enrichment {STRINGdb}	R Documentation

get_enrichment

Description

Returns the enrichment in pathways of the vector of STRING proteins that is given in input.

Usage


## S4 method for signature 'STRINGdb'
get_enrichment(string_ids, category = "Process", methodMT = "fdr", iea = TRUE, minScore=NULL)

Arguments

string_ids
a vector of STRING identifiers.

category
category for which to compute the enrichment (i.e. "Process", "Component", "Function", "KEGG", "Pfam", "InterPro"). The default category is "Process".

methodMT
method to be used for the multiple testing correction. (i.e. "fdr", "bonferroni"). The default is "fdr".

iea
specify whether you also want to use electronic inference annotations

minScore
with Tissue and Disease categories is possible to filter the annotations having an annotation score higher than this threshold (from 0 to 5)

Value

Data frame containing the enrichment in pathways of the vector of STRING proteins that is given in input.

Author(s)

Andrea Franceschini

[Package STRINGdb version 1.22.0 Index]