get_homologs {STRINGdb}	R Documentation

get_homologs

Description

Returns the homologs of the given input identifiers that are present in the given target_species_id.

Usage


## S4 method for signature 'STRINGdb'
get_homologs(string_ids, target_species_id, bitscore_threshold=NULL)

Arguments

string_ids
a vector of STRING identifiers.

target_species_id
NCBI taxonomy identifier of the species to query for homologs (the species must be present in the STRING database)

bitscore_threshold
threshold on the bitscore of the blast alignment.

Value

Data frame containing the homologs of the given input identifiers and that are present in the given target_species_id.

Author(s)

Andrea Franceschini

[Package STRINGdb version 1.22.0 Index]