get_homologs_besthits {STRINGdb}	R Documentation

get_homologs_besthits

Description

Returns the best blast hits x species of the given input identifiers.

Usage


## S4 method for signature 'STRINGdb'
get_homologs_besthits(string_ids, symbets = FALSE, target_species_id = NULL, bitscore_threshold=NULL)

Arguments

string_ids
a vector of STRING identifiers.

target_species_id
NCBI taxonomy identifier of the species to query for homologs (the species must be present in the STRING database)

bitscore_threshold
threshold on the bitscore of the blast alignment.

symbets
specify whether you want only symmetrical best hits

Value

Data frame containing the best blast hits x species of the given input identifiers.

Author(s)

Andrea Franceschini

[Package STRINGdb version 1.22.0 Index]