get_link {STRINGdb}	R Documentation

get_link

Description

Returns a short link to the network page of our STRING website that shows the protein interactions between the given identifiers.

Usage


## S4 method for signature 'STRINGdb'
get_link(string_ids, required_score=NULL, network_flavor="evidence", payload_id = NULL)

Arguments

string_ids
a vector of STRING identifiers.

required_score
minimum STRING combined score of the interactions (if left NULL we get the combined score of the object, which is 400 by default).

network_flavor
specify the flavor of the network ("evidence", "confidence" or "actions". default "evidence").

payload_id
an identifier of payload data on the STRING server (see method post_payload for additional informations)

Value

short link to the network page of our STRING website that shows the protein interactions between the input identifiers.

Author(s)

Andrea Franceschini

[Package STRINGdb version 1.22.0 Index]