get_ppi_enrichment {STRINGdb}	R Documentation

get_ppi_enrichment

Description

Returns a pvalue representing the enrichment in interactions of the list of proteins (i.e. the probability to obtain such a number of interactions by chance).

Usage


## S4 method for signature 'STRINGdb'
get_ppi_enrichment(string_ids)

Arguments

string_ids
a vector of STRING identifiers

Value

Returns a pvalue representing the enrichment in interactions of the list of proteins (i.e. the probability to obtain such a number of interactions by chance).

Author(s)

Andrea Franceschini

[Package STRINGdb version 1.22.0 Index]