get_pubmed_interaction {STRINGdb}	R Documentation

get_pubmed_interaction

Description

Returns vector with the PUBMED IDs of the publications that contain the names of both the input proteins.

Usage


## S4 method for signature 'STRINGdb'
get_pubmed_interaction(STRING_id_a, STRING_id_b )

Arguments

STRING_id_a
STRING identifier

STRING_id_b
STRING identifier

Value

vector with the PUBMED IDs of the publications that contain the names of both the input proteins.

Author(s)

Andrea Franceschini

[Package STRINGdb version 1.22.0 Index]