get_term_proteins {STRINGdb}	R Documentation

get_term_proteins

Description

Returns the proteins annotated to belong to a given term.

Usage


## S4 method for signature 'STRINGdb'
get_term_proteins(term_ids, string_ids=NULL, enableIEA=TRUE)

Arguments

term_ids
vector of terms

string_ids
a vector of STRING identifiers. If the variable is set, the method returns only the proteins that are present in this vector.

enableIEA
whether to consider also Electronic Inferred Annotations

Value

Returns the proteins annotated to belong to a given term.

Author(s)

Andrea Franceschini

[Package STRINGdb version 1.22.0 Index]