load_all {STRINGdb}	R Documentation

load_all

Description

Force download and loading of all the files (so that you can later store the object on the hard disk if you like). It makes use of the variables: "backgroundV" vector containing STRING identifiers to be used as background (i.e. the STRING network loaded will contain only the proteins that are present also in this vector) "score_threshold" STRING combined score threshold (the network loaded contains only interactions having a combined score greater than this threshold)

Usage


## S4 method for signature 'STRINGdb'
load_all()

Author(s)

Andrea Franceschini

[Package STRINGdb version 1.22.0 Index]