map {STRINGdb}	R Documentation

map

Description

Maps the gene identifiers of the input dataframe to STRING identifiers. It returns the input dataframe with the "STRING_id" additional column.

Usage


## S4 method for signature 'STRINGdb'
map(my_data_frame, my_data_frame_id_col_names, takeFirst=TRUE, removeUnmappedRows=FALSE, quiet=FALSE)

Arguments

my_data_frame
data frame provided as input.

my_data_frame_id_col_names
vector contatining the names of the columns of "my_data_frame" that have to be used for the mapping.

takeFirst
boolean indicating what to do in case of multiple STRING proteins that map to the same name. If TRUE, only the first of those is taken. Otherwise all of them are used. (default TRUE)

removeUnmappedRows
remove the rows that cannot be mapped to STRING (by default those lines are left and their STRING_id is set to NA).

quiet
Setting this variable to TRUE we can avoid printing the warning relative to the unmapped values.

Value

Returns the dataframe that is given in input with the "STRING_id" additional column.

Author(s)

Andrea Franceschini

[Package STRINGdb version 1.22.0 Index]