mp {STRINGdb}	R Documentation

mp

Description

Maps the gene identifiers of the input vector to STRING identifiers (using a take first approach). It returns a vector with the STRING identifiers of the mapped proteins.

Usage


## S4 method for signature 'STRINGdb'
mp(protein_aliases)

Arguments

protein_aliases
vector of protein aliases that we want to convert to STRING identifiers

Value

It returns a vector with the STRING identifiers of the mapped proteins.

Author(s)

Andrea Franceschini

[Package STRINGdb version 1.22.0 Index]