multi_map_df {STRINGdb}	R Documentation

multi_map_df

Description

mapping function (it add the possibility to map using more than one column of the data frame)

Usage


multi_map_df(dfToMap, dfMap, strColsFrom, strColFromDfMap, strColToDfMap, caseSensitive=FALSE)

Arguments

dfToMap
input data frame (that contains the columns that need to be mapped)

dfMap
data frame containing the mapping data

strColsFrom
sorted vector containing the names of the columns to be used in the input data frame for the mapping (the order of the elements in the vector defines the priority for the mapping)

strColFromDfMap
name of the column in the mapping data frame to be used as source for the mapping

strColToDfMap
name of the column in the mapping data frame to be used as target for the mapping

caseSensitive
specify whether the mapping should be case sensitive

Value

data frame with an additional column containing the result of the mapping

Author(s)

Andrea Franceschini

[Package STRINGdb version 1.22.0 Index]