plot_network {STRINGdb}	R Documentation

plot_network

Description

Plots an image of the STRING network with the given proteins.

Usage


## S4 method for signature 'STRINGdb'
plot_network(string_ids, payload_id=NULL, required_score=NULL, add_link=TRUE, add_summary=TRUE)

Arguments

string_ids
a vector of STRING identifiers

payload_id
an identifier of payload data on the STRING server (see method post_payload for additional informations)

required_score
a threshold on the score that overrides the default score_threshold, that we use only for the picture

add_link
parameter to specify whether you want to generate and add a short link to the relative page in STRING. As default this option is active but we suggest to deactivate it in case one is generating many images (e.g. in a loop). Deactivating this option avoids to generate and store a lot of short-urls on our server.

add_summary
parameter to specify whether you want to add a summary text to the picture. This summary includes a p-value and the number of proteins/interactions.

Author(s)

Andrea Franceschini

[Package STRINGdb version 1.22.0 Index]