plot_ppi_enrichment {STRINGdb}	R Documentation

plot_ppi_enrichment

Description

Plots a graph showing the enrichment in protein interactions in various positions of the list of genes in input.

Usage


## S4 method for signature 'STRINGdb'
plot_ppi_enrichment(string_ids, file=NULL, sliceWindow = 20, edgeWindow = 140, 
                                     windowExtendedReferenceThreshold = 260, minVal=0.0000000001, title="", quiet=FALSE)

Arguments

string_ids
a vector of STRING identifiers

file
file where to save the graph as an image

sliceWindow
defines the interval in proteins after which to compute the enrichment, scanning the list (i.e. the resolution)

edgeWindow
size of the window that we use to compute the enrichment (i.e. the window pvalue is computed using the proteins inside this "edgeWindow")

windowExtendedReferenceThreshold
defines the size of a window at the beginning of the list. The enrichment will be computed always including the proteins in this window

title
title of the graph.

minVal
minimum value that the pvalue can assume in the log-scale graph. If the p-value is lower, we convert the value to this minimum value

quiet
if set to TRUE the method runs in quiet mode (turning off any output message)

Author(s)

Andrea Franceschini

[Package STRINGdb version 1.22.0 Index]