post_payload {STRINGdb}	R Documentation

post_payload

Description

Posts the input to STRING and returns an identifier that you can use to access the payload when you enter in our website.

Usage


## S4 method for signature 'STRINGdb'
post_payload(stringIds, colors=NULL, comments=NULL, links=NULL, iframe_urls=NULL, logo_imgF=NULL, legend_imgF=NULL )

Arguments

stringIds
vector of STRING identifiers.

colors
vector containing the colors to use for a every STRING identifier ( the order of the elements must match those in the string_ids vector)

comments
vector containing the comments to use for every STRING identifier ( the order of the elements must match those in the string_ids vector)

links
vector containing the links to use for every STRING identifier ( the order of the elements must match those in the string_ids vector)

iframe_urls
vector containing the urls of the iframes to use for every STRING identifier ( the order of the elements must match those in the string_ids vector).

logo_imgF
path to a file containing the logo image to be display in the STRING website

legend_imgF
path to a file containing a legend image to be display in the STRING website

Value

identifier of the payload.

Author(s)

Andrea Franceschini

[Package STRINGdb version 1.22.0 Index]