ppie.compLambdaL1L2 {STRINGdb}	R Documentation

ppie.compLambdaL1L2

Description

Compute lambda L1L2

Usage


ppie.compLambdaL1L2(degreesI, degreesJ, edgeNum)

Arguments

degreesI
vector containing the degrees of the nodes present in nodeSet I

degreesJ
vector containing the degrees of the nodes present in nodeSet J

edgeNum
total number of edges of the entire graph

Details

compute the number of expected interactions between two sets of nodes

Author(s)

Andrea Franceschini

[Package STRINGdb version 1.22.0 Index]