ppie.compPij {STRINGdb}	R Documentation

ppie.compPij

Description

Compute the pvalue of protein I to interact with protein J (look at the Pradines paper referenced below)

Usage


ppie.compPij(degI, degJ, edgeNum)

Arguments

degI
degree of protein I

degJ
degree of protein J

edgeNum
total number of edges of the entire graph

Value

return the pvalue of protein I to interact with protein J (look at the Pradines paper referenced below)

Author(s)

Andrea Franceschini

References

Pradines JR, Farutin V, Rowley S, Dancik V. Analyzing protein lists with large networks: edge-count probabilities in random graphs with given expected degrees. J. Comput. Biol. 2005;12:113-128. Franceschini, A et al. (2013). STRING v9.1: protein-protein interaction networks, with increased coverage and integration. In:Nucleic Acids Res. 2013 Jan;41(Database issue)

[Package STRINGdb version 1.22.0 Index]