ppie.getNumEdgesBetween {STRINGdb}	R Documentation

ppie.getNumEdgesBetween

Description

find the number of interactions between two sets of nodes

Usage


ppie.getNumEdgesBetween(graph, nodesFrom, nodesTo)

Arguments

graph
igraph object

nodesFrom
list of nodes

nodesTo
list of nodes

Value

return the number of interactions between two sets of nodes

Author(s)

Andrea Franceschini

[Package STRINGdb version 1.22.0 Index]