set_background {STRINGdb}	R Documentation

set_background

Description

With this method you can specify a vector of proteins to be used as background. The network is reloaded and only the proteins that are present in the background vector are inserted in the graph. Besides, the background is taken in consideration for all the enrichment statistics.

Usage


## S4 method for signature 'STRINGdb'
set_background(background_vector )

Arguments

background_vector
vector of STRING protein identifiers

Author(s)

Andrea Franceschini

[Package STRINGdb version 1.22.0 Index]